Fixed pointer type mismatch

+++ src/bootloader/stage2/x86.h
@@ -2,6 +2,6 @@
 
 #include "stdint.h"
 
-void _cdecl x86_div64_32(uint64_t dividend, uint32_t divisor, uint32_t* quotientOut, uint32_t* remainderOut);
+void _cdecl x86_div64_32(uint64_t dividend, uint32_t divisor, uint64_t* quotientOut, uint32_t* remainderOut);
 
 void _cdecl x86_VideoWriteCharTeletype(char c, uint8_t page);
